refactor: move logic implementation to vaults store
